% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nps_price.R
\name{get_NPS_price}
\alias{get_NPS_price}
\title{Get NPS price data}
\usage{
get_NPS_price(query_start = "2019-01-01 00:00",
  query_end = "2019-01-02 00:00")
}
\description{
Get NPS price data
}
