% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_system_plan.R
\name{get_system_plan}
\alias{get_system_plan}
\title{Get System plan data}
\usage{
get_system_plan(query_start = "2019-01-01 00:00",
  query_end = "2019-01-02 00:00")
}
\description{
Get System plan data
}
