% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_system_real.R
\name{get_system_real}
\alias{get_system_real}
\title{Get System real data}
\usage{
get_system_real(query_start = "2019-01-01 00:00",
  query_end = "2019-01-02 00:00")
}
\description{
Get System real data
}
